backend/materials_project/requirements.txt
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.1
cached-property==1.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
distro==1.8.0
Django==3.2.7
django-colorfield==0.7.2
django-cors-headers==3.9.0
django-filter==2.4.0
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-extra-fields==3.4.1
drf-yasg==1.20
flake8==5.0.4
geojson==2.5.0
gunicorn==20.0.4
idna==3.4
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==21.3
paramiko==2.12.0
Pillow==9.3.0
psycopg2-binary==2.8.6
py-msyh==0.0.0.3
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.1.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2020.1
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.3.1
texttable==1.6.7
tomli==2.0.1
ttf==0.1.0
tweepy==4.12.1
typed-ast==1.5.4
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
websocket-client==0.59.0
zipp==3.10.0
daphne==4.0
channels
channels-redis
